requirements.txt
faiss-cpu
groq
langchain-groq
pyPDF2
django
langchain
langchain_community
langchain_google_genai
python-dotenv
pypdf
langchain_huggingface
sentencepiece
 accelerate>=0.26.0
